Some more conversions to LONG.

+++ src/hddEditor/libs/disks/Disk.java
@@ -69,7 +69,7 @@ public interface Disk {
 	 * @return
 	 * @throws IOException
 	 */
-	public byte[] GetBytesStartingFromSector(long SectorNum, int sz) throws IOException;
+	public byte[] GetBytesStartingFromSector(long SectorNum, long sz) throws IOException;
 
 	/**
 	 *
