+++ src/hddEditor/libs/partitions/CPMPartition.java
@@ -78,17 +78,13 @@ public class CPMPartition extends IDEDosPartition {
 	 */
 	public byte[] GetLogicalBlock(int BlockID) throws IOException {
 
-		int LogicalSectorInPartition = (BlockID * BlockSize) / CurrentDisk.GetSectorSize();
+		long LogicalSectorInPartition = (BlockID * BlockSize) / CurrentDisk.GetSectorSize();
 
-		int LogicalSectorStartOfPartition = (GetStartCyl() + ReservedTracks) * CurrentDisk.GetNumHeads()
+		long LogicalSectorStartOfPartition = (GetStartCyl() + ReservedTracks) * CurrentDisk.GetNumHeads()
 				* (CurrentDisk.GetNumSectors());
 		LogicalSectorStartOfPartition = LogicalSectorStartOfPartition + (GetStartHead() * CurrentDisk.GetNumSectors());
 
-		int ActualLogicalSector = LogicalSectorStartOfPartition + LogicalSectorInPartition;
-
-//		System.out.println("GetLogicalBlock: " + BlockID + " SC:" + GetStartCyl() + " SH:" + GetStartHead()
-//		+ " RealSector:" + ActualLogicalSector + " Log:" + LogicalSectorInPartition + " Startsect:"
-//		+ LogicalSectorStartOfPartition);
+		long ActualLogicalSector = LogicalSectorStartOfPartition + LogicalSectorInPartition;
 
 		byte result[] = CurrentDisk.GetBytesStartingFromSector(ActualLogicalSector, BlockSize);
